Next step on NRF24L01P socket device driver; Connection-oriented Client/Server.
1. Adding initial support for connection-oriented sockets (Client/Server).
2. Adding example sketches; CosaNRFclient/CosaNRFserver.

+++ Cosa/Socket.hh
@@ -75,7 +75,7 @@ public:
      * @param[in] s socket to attach.
      * @return binding or negative error code.
      */
-    int attach(Socket* s);
+    int8_t attach(Socket* s);
 
     /**
      * Detach given socket from device.
